requirements.txt
Django==2.2.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.1.1
pytz==2019.3
sqlparse==0.3.0
